fix renderer
