+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kw-blog-nuxt",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "private": true,
   "type": "module",
   "scripts": {
@@ -8,12 +8,14 @@
     "dev": "nuxt dev",
     "generate": "nuxt generate",
     "preview": "nuxt preview",
-    "postinstall": "nuxt prepare"
+    "postinstall": "nuxt prepare",
+    "aws:deploy": "serverless deploy"
   },
   "devDependencies": {
     "@nuxt/test-utils": "^3.9.0",
     "@nuxtjs/tailwindcss": "^6.10.4",
     "nuxt": "^3.9.0",
+    "serverless": "^3.38.0",
     "tailwindcss": "^3.4.1",
     "vue": "^3.4.6",
     "vue-router": "^4.2.5"
