+++ utils/renderer.ts
@@ -1,208 +1,16 @@
-import { h } from "vue";
 import type { VNode } from "vue";
-import { IBLOCKS } from "types";
-import type {
-  IAnnotations,
-  IRichText,
-  IRichTextProps,
-  IBlock,
-  IRenderer,
-} from "types";
+import type { IBlock, BlockRendererProps } from "types";
+import { defaultBlockRenderers } from "./defaultBlockRenderer";
+import { defaultRichTextRenderer } from "./defaultRichTextRenderer";
 
-export const defaultFontStyle = {
-  "font-family": `ui-sans-serif, -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, "Apple Color Emoji", Arial, sans-serif, "Segoe UI Emoji", "Segoe UI Symbol"`,
-  "caret-color": "rgb(55, 53, 47)",
-};
-
-function handleAnnotationStyles(annotation: IAnnotations) {
-  const styles = {
-    ...defaultFontStyle,
-  };
-  if (annotation.bold) {
-    styles["font-weight"] = "bold";
-  }
-  if (annotation.italic) {
-    styles["font-style"] = "italic";
-  }
-  if (annotation.underline) {
-    styles["text-decoration"] = "underline";
-  }
-  if (annotation.strikethrough) {
-    styles["text-decoration"] = "line-through";
-  }
-  if (annotation.code) {
-    styles["font-family"] = "monospace";
-  }
-  if (annotation.color) {
-    styles["color"] = annotation.color;
-  }
-  return styles;
-}
-
-function defaultRichTextRenderer(richTextArray: IRichText[], key: string) {
-  if (!richTextArray || !richTextArray.length) {
-    return null;
-  }
-  const result = richTextArray.map((richText, i) => {
-    const attributes = {
-      style: handleAnnotationStyles(richText.annotations),
-      key: `${key}-annotation-${i}`,
-    };
-    if (richText.text.link) attributes["href"] = richText.href;
-    if (i < richTextArray.length - 1)
-      attributes["style"]["margin-right"] = "4px";
-    return h(
-      richText.text.link ? "a" : "span",
-      attributes,
-      richText.text.content
-    );
+function renderBlocks(blocks: IBlock[], renderer: any) {
+  return blocks.map((block, i) => {
+    return renderBlock(block, renderer);
   });
-  return h("span", { key }, result);
 }
 
-const defaultBlockRenderers = {
-  [IBLOCKS.heading_1]: (block: IBlock, key: string, next: VNode) => {
-    return h(
-      "h1",
-      {
-        key,
-        style: {
-          "font-size": "1.875em",
-        },
-      },
-      next
-    );
-  },
-  [IBLOCKS.heading_2]: (block: IBlock, key: string, next: VNode) => {
-    return h(
-      "h2",
-      {
-        key,
-        style: {
-          "font-size": "1.5em",
-        },
-      },
-      next
-    );
-  },
-  [IBLOCKS.heading_3]: (block: IBlock, key: string, next: VNode) => {
-    return h(
-      "h3",
-      {
-        key,
-        style: {
-          "font-size": "1.25em",
-        },
-      },
-      next
-    );
-  },
-  [IBLOCKS.paragraph]: (block: IBlock, key: string, next: VNode) => {
-    return h(
-      "p",
-      {
-        key,
-        style: {
-          "font-size": "1em",
-          "word-wrap": "break-word",
-          "white-space": "pre-wrap",
-        },
-      },
-      next
-    );
-  },
-  [IBLOCKS.bulleted_list_item]: (block: IBlock, key: string, next: VNode) => {
-    const childeLi = h("li", { key, style: { padding: "0.25em  0" } }, next);
-    return h(
-      "ul",
-      {
-        key,
-        style: {},
-      },
-      childeLi
-    );
-  },
-  [IBLOCKS.numbered_list_item]: (block: IBlock, key: string, next: VNode) => {
-    const childeLi = h("li", { key, style: { padding: "0.25em  0" } }, next);
-    return h(
-      "ol",
-      {
-        key,
-        style: {},
-      },
-      childeLi
-    );
-  },
-  [IBLOCKS.to_do]: (block: IBlock, key: string, next: VNode) => {
-    const checkBox = h("input", {
-      type: "checkbox",
-      style: {
-        boxSizing: "border-box",
-        height: "16px",
-        width: "16px",
-        border: block.to_do?.checked ? "none" : "1px solid black",
-        backgroundColor: block.to_do?.checked ? "rgb(35, 131, 226)" : "white",
-      },
-      checked: block.to_do?.checked,
-    });
-    const checkBoxTextStyle = {
-      textDecoration: block.to_do?.checked
-        ? "line-through rgba(55, 53, 47, 0.25)"
-        : "none",
-      color: block.to_do?.checked ? "rgba(55, 53, 47, 0.65)" : "inherit",
-    };
-    const checkBoxTextArea = h(
-      "div",
-      { key, style: { ...checkBoxTextStyle, padding: "0.25em  0" } },
-      next
-    );
-    return h(
-      "div",
-      {
-        key,
-        style: {
-          display: "flex",
-          gap: "0.5em",
-          alignItems: "center",
-        },
-      },
-      [checkBox, checkBoxTextArea]
-    );
-  },
-  [IBLOCKS.divider]: (block, key, next) => {
-    return h("hr", {
-      key,
-      style: { width: "100%", color: "rgba(55, 53, 47, 0.16)" },
-    });
-  },
-  text: (richTextArray: IRichText[], key, annotationRenderer) => {
-    if (!richTextArray.length) {
-      return null;
-    }
-    const result = richTextArray.map((richText, i) => {
-      h(
-        "span",
-        {
-          style: handleAnnotationStyles(richText.annotations),
-          key: `annotation-${i}`,
-        },
-        richText.text.content
-      );
-    });
-    return h("span", { key }, result);
-  },
-};
-
-function renderBlockList(
-  blocks: IBlock[],
-  key: string,
-  renderer: IRenderer
-): VNode[] {
-  return blocks.map((block, i) => renderBlock(block, `${key}-${i}`, renderer));
-}
-
-function renderBlock(block: IBlock, key: string, renderer: IRenderer) {
-  const blockRenderer = renderer.block;
+function renderBlock(block: IBlock, renderer: any) {
+  const blockRenderer = renderer;
   if (!blockRenderer[block.type]) {
     console.warn(
       `[notion-vue-renderer] Unsupported block type "${block.type}". Please create an issue if you'd like to see it supported`
@@ -211,47 +19,23 @@ function renderBlock(block: IBlock, key: string, renderer: IRenderer) {
   }
   return blockRenderer[block.type](
     block,
-    key,
-    defaultRichTextRenderer(block[block.type].rich_text, key) as VNode
+    defaultRichTextRenderer(block[block.type].rich_text)
   );
 }
-
-const RichText = ({
-  blockRenderers,
-  annotationRenderers,
-  blocks,
-}: IRichTextProps) => {
-  if (!blocks) {
-    return console.warn("[notion-vue-renderer] No blocks provided");
-  }
-  const renderer = {
-    block: {
-      ...defaultBlockRenderers,
-    },
-  } as any;
-  if (blockRenderers) {
-    renderer.block = {
-      ...renderer.block,
-      ...blockRenderers,
-    };
-  }
-  const results = renderBlockList(blocks, "block", renderer);
-  return formatLists(removeNulls(results));
-};
 
 function removeNulls(array: VNode[]) {
   return array.filter((item) => item !== null);
 }
 
-function formatLists(blocks: any[]): VNode[] {
-  const output: any[] = [];
+function formatLists(blocks: VNode[]): VNode[] {
+  const output: VNode[] = [];
   const indexesToRemove: any[] = [];
   let lastBlock: VNode | null = null;
   for (const block of blocks) {
     if (lastBlock && block.type === "ul" && lastBlock.type === "ul") {
       const newUnorderedList = h("ul", { key: lastBlock.key as any }, [
         ...(lastBlock.children as VNode[]),
-        ...block.children,
+        ...(block.children as VNode[]),
       ]);
       output.push(newUnorderedList);
       indexesToRemove.push(output.length - 2);
@@ -260,7 +44,7 @@ function formatLists(blocks: any[]): VNode[] {
       const newOrderedList = h(
         "ol",
         { key: lastBlock.key as any, style: { ...defaultFontStyle } },
-        [...(lastBlock.children as any), ...block.children]
+        [...(lastBlock.children as any), ...(block.children as VNode[])]
       );
       output.push(newOrderedList);
       indexesToRemove.push(output.length - 2);
@@ -273,6 +57,14 @@ function formatLists(blocks: any[]): VNode[] {
   return output.filter((item, index) => !indexesToRemove.includes(index));
 }
 
-RichText.props = ["blocks", "blockRenderers", "annotationRenderers"];
+const RichTexts = ({ blocks }: BlockRendererProps): VNode[] | void => {
+  if (!blocks) {
+    return console.warn("[notion-vue-renderer] No blocks provided");
+  }
+  const results = renderBlocks(blocks, defaultBlockRenderers);
+  return removeNulls(results);
+};
 
-export default RichText;
+RichTexts.props = ["blocks"];
+
+export default RichTexts;
